requirements.txt
streamlit
biopython
langchain
faiss-cpu
sentence-transformers
google-generativeai